add UDS server for sending read data other minor adjustments

+++ server/SensorReading.hpp
@@ -1,12 +1,11 @@
 #pragma once
 #include <nlohmann/json.hpp>
-#include <nlohmann/json_fwd.hpp>
 struct SensorReading {
   float value;
   float min_value;
   float max_value;
   double sum;
-  unsigned int times;
+  size_t times;
 
   nlohmann::json Serialize() {
     nlohmann::json j;
